fix: update SDK to use builtInStyles from core package [ALT-227]
