+++ packages/core/src/definitions/styles.ts
@@ -282,3 +282,14 @@ export const optionalBuiltInStyles: Partial<
     defaultValue: false,
   },
 };
+
+export const containerBuiltInStyles = {
+  ...builtInStyles,
+  cfHeight: {
+    displayName: 'Height',
+    type: 'Text',
+    group: 'style',
+    description: 'The height of the section',
+    defaultValue: 'auto',
+  } as ComponentDefinitionVariable<'Text'>,
+};
